requirements.txt
APScheduler==3.7.0
discord.py==2.1.0
discord_webhook==0.17.0
firebase_admin==5.3.0
google_api_python_client==2.74.0
hikari==2.0.0.dev115
hikari_lightbulb==2.3.1
python-dotenv==0.21.1
pytz==2022.4
requests==2.27.1
